dnsdist: Add unit tests for the load-balancing policies
