+++ pdns/dnsdistdist/dnsdist-lbpolicies.hh
@@ -1,0 +1,1 @@
+../dnsdist-lbpolicies.hh
